chore: Move proxies to their own packages
