+++ plugins/node/opentelemetry-instrumentation-pulsar-client/src/index.ts
@@ -20,3 +20,6 @@ export * from './types';
 export * from './instrumentation';
 export default Instrumentation;
 export { Instrumentation as PulsarInstrumentation };
+export {ConsumerProxy} from "./proxies/consumer";
+export {ProducerProxy} from "./proxies/producer";
+export {ClientProxy} from "./proxies/client";
